+++ crates/bonsaidb/src/any_connection.rs
@@ -256,7 +256,7 @@ impl<B: Backend> Connection for AnyDatabase<B> {
         order: Sort,
         limit: Option<usize>,
         access_policy: AccessPolicy,
-    ) -> Result<Vec<MappedDocument<V::Key, V::Value>>, bonsaidb_core::Error>
+    ) -> Result<Vec<MappedDocument<V>>, bonsaidb_core::Error>
     where
         Self: Sized,
     {